10_demo_day_project/requirements.txt
# Core dependencies
tensorflow
numpy
pandas
matplotlib

# Backend API & Deployment
mlflow
fastapi
uvicorn

# Frontend (Streamlit app)
streamlit

# AWS interactions
boto3

# Image processing (needed for handling skin images)
Pillow

# Machine Learning utilities
scikit-learn

# FastAPI = gestion des fichiers envoyés
python-multipart